+++ custom_components/jarvis/cognitive_core.py
@@ -2181,7 +2181,8 @@ async def _tick():
                 preds = (cognition.predict(hass, now_t)
                          + cognition.predict_overdue(hass, now_t)
                          + cognition.predict_presence(hass, now_t)
-                         + cognition.predict_proximity(hass, now_t))
+                         + cognition.predict_proximity(hass, now_t)
+                         + cognition.predict_departure(hass, now_t))
                 for pred in preds:
                     actions.append(pred)
                     from .websocket import jarvis_log